+++ services/hh/queue.go
@@ -72,6 +72,10 @@ type queue struct {
 	// The segments that exist on disk
 	segments segments
 }
+type queuePos struct {
+	head string
+	tail string
+}
 
 type segments []*segment
 
@@ -211,7 +215,21 @@ func (l *queue) LastModified() (time.Time, error) {
 	if l.tail != nil {
 		return l.tail.lastModified()
 	}
-	return time.Time{}, nil
+	return time.Time{}.UTC(), nil
+}
+
+func (l *queue) Position() (*queuePos, error) {
+	l.mu.RLock()
+	defer l.mu.RUnlock()
+
+	qp := &queuePos{}
+	if l.head != nil {
+		qp.head = fmt.Sprintf("%s:%d", l.head.path, l.head.pos)
+	}
+	if l.tail != nil {
+		qp.tail = fmt.Sprintf("%s:%d", l.tail.path, l.tail.filePos())
+	}
+	return qp, nil
 }
 
 // diskUsage returns the total size on disk used by the queue
@@ -606,7 +624,7 @@ func (l *segment) lastModified() (time.Time, error) {
 	if err != nil {
 		return time.Time{}, err
 	}
-	return stats.ModTime(), nil
+	return stats.ModTime().UTC(), nil
 }
 
 func (l *segment) diskUsage() int64 {