+++ services/hh/service.go
@@ -3,9 +3,11 @@ package hh
 import (
 	"expvar"
 	"fmt"
-	"io"
+	"io/ioutil"
 	"log"
 	"os"
+	"path/filepath"
+	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -13,15 +15,17 @@ import (
 	"github.com/influxdb/influxdb"
 	"github.com/influxdb/influxdb/meta"
 	"github.com/influxdb/influxdb/models"
+	"github.com/influxdb/influxdb/monitor"
 )
 
 var ErrHintedHandoffDisabled = fmt.Errorf("hinted handoff disabled")
 
 const (
-	writeShardReq       = "write_shard_req"
-	writeShardReqPoints = "write_shard_req_points"
-	processReq          = "process_req"
-	processReqFail      = "process_req_fail"
+	writeShardReq       = "wr_shard_req"
+	writeShardReqPoints = "wr_shard_req_points"
+	writeNodeReq        = "wr_node_req"
+	writeNodeReqFail    = "wr_node_req_fail"
+	writeNodeReqPoints  = "wr_node_req_points"
 )
 
 type Service struct {
@@ -29,17 +33,18 @@ type Service struct {
 	wg      sync.WaitGroup
 	closing chan struct{}
 
+	processors map[uint64]*NodeProcessor
+
 	statMap *expvar.Map
 	Logger  *log.Logger
 	cfg     Config
 
-	ShardWriter shardWriter
+	shardWriter shardWriter
+	metastore   metaStore
 
-	HintedHandoff interface {
-		WriteShard(shardID, ownerID uint64, points []models.Point) error
-		Process() error
-		PurgeOlderThan(when time.Duration) error
-		PurgeInactiveOlderThan(when time.Duration) error
+	Monitor interface {
+		RegisterDiagnosticsClient(name string, client monitor.DiagsClient)
+		DeregisterDiagnosticsClient(name string)
 	}
 }
 
@@ -56,55 +61,81 @@ func NewService(c Config, w shardWriter, m metaStore) *Service {
 	key := strings.Join([]string{"hh", c.Dir}, ":")
 	tags := map[string]string{"path": c.Dir}
 
-	s := &Service{
-		cfg:     c,
-		statMap: influxdb.NewStatistics(key, "hh", tags),
-		Logger:  log.New(os.Stderr, "[handoff] ", log.LstdFlags),
+	return &Service{
+		cfg:         c,
+		closing:     make(chan struct{}),
+		processors:  make(map[uint64]*NodeProcessor),
+		statMap:     influxdb.NewStatistics(key, "hh", tags),
+		Logger:      log.New(os.Stderr, "[handoff] ", log.LstdFlags),
+		shardWriter: w,
+		metastore:   m,
 	}
-	processor, err := NewProcessor(c.Dir, w, m, ProcessorOptions{
-		MaxSize:        c.MaxSize,
-		RetryRateLimit: c.RetryRateLimit,
-	})
-	if err != nil {
-		s.Logger.Fatalf("Failed to start hinted handoff processor: %v", err)
-	}
-
-	processor.Logger = s.Logger
-	s.HintedHandoff = processor
-	return s
 }
 
 func (s *Service) Open() error {
-	if !s.cfg.Enabled {
-		// Allow Open to proceed, but don't anything.
-		return nil
-	}
-
-	s.Logger.Printf("Starting hinted handoff service")
-
 	s.mu.Lock()
 	defer s.mu.Unlock()
-
+	if !s.cfg.Enabled {
+		// Allow Open to proceed, but don't do anything.
+		return nil
+	}
+	s.Logger.Printf("Starting hinted handoff service")
 	s.closing = make(chan struct{})
 
-	s.Logger.Printf("Using data dir: %v", s.cfg.Dir)
+	// Register diagnostics if a Monitor service is available.
+	if s.Monitor != nil {
+		s.Monitor.RegisterDiagnosticsClient("hh", s)
+	}
 
-	s.wg.Add(3)
-	go s.retryWrites()
-	go s.expireWrites()
-	go s.deleteInactiveQueues()
+	// Create the root directory if it doesn't already exist.
+	s.Logger.Printf("Using data dir: %v", s.cfg.Dir)
+	if err := os.MkdirAll(s.cfg.Dir, 0700); err != nil {
+		return fmt.Errorf("mkdir all: %s", err)
+	}
+
+	// Create a node processor for each node directory.
+	files, err := ioutil.ReadDir(s.cfg.Dir)
+	if err != nil {
+		return err
+	}
+
+	for _, file := range files {
+		nodeID, err := strconv.ParseUint(file.Name(), 10, 64)
+		if err != nil {
+			// Not a number? Skip it.
+			continue
+		}
+
+		n := NewNodeProcessor(nodeID, s.pathforNode(nodeID), s.shardWriter, s.metastore)
+		if err := n.Open(); err != nil {
+			return err
+		}
+		s.processors[nodeID] = n
+	}
+
+	s.wg.Add(1)
+	go s.purgeInactiveProcessors()
 
 	return nil
 }
 
 func (s *Service) Close() error {
+	s.Logger.Println("shutting down hh service")
 	s.mu.Lock()
 	defer s.mu.Unlock()
+
+	for _, p := range s.processors {
+		if err := p.Close(); err != nil {
+			return err
+		}
+	}
 
 	if s.closing != nil {
 		close(s.closing)
 	}
 	s.wg.Wait()
+	s.closing = nil
+
 	return nil
 }
 
@@ -115,76 +146,125 @@ func (s *Service) SetLogger(l *log.Logger) {
 
 // WriteShard queues the points write for shardID to node ownerID to handoff queue
 func (s *Service) WriteShard(shardID, ownerID uint64, points []models.Point) error {
-	s.statMap.Add(writeShardReq, 1)
-	s.statMap.Add(writeShardReqPoints, int64(len(points)))
 	if !s.cfg.Enabled {
 		return ErrHintedHandoffDisabled
 	}
+	s.statMap.Add(writeShardReq, 1)
+	s.statMap.Add(writeShardReqPoints, int64(len(points)))
 
-	return s.HintedHandoff.WriteShard(shardID, ownerID, points)
-}
+	s.mu.RLock()
+	processor, ok := s.processors[ownerID]
+	s.mu.RUnlock()
+	if !ok {
+		if err := func() error {
+			// Check again under write-lock.
+			s.mu.Lock()
+			defer s.mu.Unlock()
 
-func (s *Service) retryWrites() {
-	defer s.wg.Done()
-	currInterval := time.Duration(s.cfg.RetryInterval)
-	if currInterval > time.Duration(s.cfg.RetryMaxInterval) {
-		currInterval = time.Duration(s.cfg.RetryMaxInterval)
-	}
-
-	for {
-
-		select {
-		case <-s.closing:
-			return
-		case <-time.After(currInterval):
-			s.statMap.Add(processReq, 1)
-			if err := s.HintedHandoff.Process(); err != nil && err != io.EOF {
-				s.statMap.Add(processReqFail, 1)
-				s.Logger.Printf("retried write failed: %v", err)
-
-				currInterval = currInterval * 2
-				if currInterval > time.Duration(s.cfg.RetryMaxInterval) {
-					currInterval = time.Duration(s.cfg.RetryMaxInterval)
+			processor, ok = s.processors[ownerID]
+			if !ok {
+				processor = NewNodeProcessor(ownerID, s.pathforNode(ownerID), s.shardWriter, s.metastore)
+				if err := processor.Open(); err != nil {
+					return err
 				}
-			} else {
-				// Success! Return to configured interval.
-				currInterval = time.Duration(s.cfg.RetryInterval)
+				s.processors[ownerID] = processor
 			}
+			return nil
+		}(); err != nil {
+			return err
 		}
 	}
+
+	if err := processor.WriteShard(shardID, points); err != nil {
+		return err
+	}
+
+	return nil
 }
 
-// expireWrites will cause the handoff queues to remove writes that are older
-// than the configured threshold
-func (s *Service) expireWrites() {
+// Diagnostics returns diagnostic information.
+func (s *Service) Diagnostics() (*monitor.Diagnostic, error) {
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+
+	d := &monitor.Diagnostic{
+		Columns: []string{"node", "active", "last modified", "head", "tail"},
+		Rows:    make([][]interface{}, 0, len(s.processors)),
+	}
+
+	for k, v := range s.processors {
+		lm, err := v.LastModified()
+		if err != nil {
+			return nil, err
+		}
+
+		active := "no"
+		b, err := v.Active()
+		if err != nil {
+			return nil, err
+		}
+		if b {
+			active = "yes"
+		}
+
+		d.Rows = append(d.Rows, []interface{}{k, active, lm, v.Head(), v.Tail()})
+	}
+	return d, nil
+}
+
+// purgeInactiveProcessors will cause the service to remove processors for inactive nodes.
+func (s *Service) purgeInactiveProcessors() {
 	defer s.wg.Done()
-	ticker := time.NewTicker(time.Hour)
+	ticker := time.NewTicker(time.Duration(s.cfg.PurgeInterval))
 	defer ticker.Stop()
+
 	for {
 		select {
 		case <-s.closing:
 			return
 		case <-ticker.C:
-			if err := s.HintedHandoff.PurgeOlderThan(time.Duration(s.cfg.MaxAge)); err != nil {
-				s.Logger.Printf("purge write failed: %v", err)
-			}
+			func() {
+				s.mu.Lock()
+				defer s.mu.Unlock()
+
+				for k, v := range s.processors {
+					lm, err := v.LastModified()
+					if err != nil {
+						s.Logger.Printf("failed to determine LastModified for processor %d: %s", k, err.Error())
+						continue
+					}
+
+					active, err := v.Active()
+					if err != nil {
+						s.Logger.Printf("failed to determine if node %d is active: %s", k, err.Error())
+						continue
+					}
+					if active {
+						// Node is active.
+						continue
+					}
+
+					if !lm.Before(time.Now().Add(-time.Duration(s.cfg.MaxAge))) {
+						// Node processor contains too-young data.
+						continue
+					}
+
+					if err := v.Close(); err != nil {
+						s.Logger.Printf("failed to close node processor %d: %s", k, err.Error())
+						continue
+					}
+					if err := v.Purge(); err != nil {
+						s.Logger.Printf("failed to purge node processor %d: %s", k, err.Error())
+						continue
+					}
+					delete(s.processors, k)
+				}
+			}()
 		}
 	}
 }
 
-// deleteInactiveQueues will cause the service to remove queues for inactive nodes.
-func (s *Service) deleteInactiveQueues() {
-	defer s.wg.Done()
-	ticker := time.NewTicker(time.Hour)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-s.closing:
-			return
-		case <-ticker.C:
-			if err := s.HintedHandoff.PurgeInactiveOlderThan(time.Duration(s.cfg.MaxAge)); err != nil {
-				s.Logger.Printf("delete queues failed: %v", err)
-			}
-		}
-	}
+// pathforNode returns the directory for HH data, for the given node.
+func (s *Service) pathforNode(nodeID uint64) string {
+	return filepath.Join(s.cfg.Dir, fmt.Sprintf("%d", nodeID))
 }